software/src_python/requirements-parapred.txt
pandas==2.2.3
numpy==2.2.6
polars-lts-cpu==1.33.1
polars-ds-lts-cpu==0.10.2
torch==2.7.0
click==8.1.8
parapred-pytorch
